+++ fydp-web/src/main/java/com/thomasdedinsky/fydp/fydpweb/data/ECGValueRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ECGValueRepository extends JpaRepository<ECGValue, Integer> {
+public interface ECGValueRepository extends JpaRepository<ECGValue, Long> {
     List<ECGValue> findByWristband(Wristband wristband);
     Page<ECGValue> findByWristband(Wristband wristband, Pageable pageable);
 }
