+++ fydp-web/src/main/java/com/thomasdedinsky/fydp/fydpweb/data/HubRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface HubRepository extends JpaRepository<Hub, Integer> {
+public interface HubRepository extends JpaRepository<Hub, Long> {
     List<Hub> findByIdNotIn(List<Long> idList);
 }
